Move ProgressBar para o template
